habesha/requirements.txt
rembg==2.0.65
Pillow 